refactor(go): rename models to domain
