+++ backend/internal/domain/permission.go
@@ -1,4 +1,4 @@
-package models
+package domain
 
 import "strings"
