+++ backend/internal/domain/permission_test.go
@@ -1,4 +1,4 @@
-package models
+package domain
 
 import (
 	"fmt"
